Refactor livingroom into generic 'zone' framework
There is now a base 'zone' class that is used as a framework for individuali
thermostats.

This new code introduces secondary heating and fan control as well.  Since
it is still summer here, there won't be much testing on this code for a
while still.

The zone contains all of the generic algorithms for heating (secondary),
cooling and fan control.  Each of these operations can be replace with
zone specific logic, if needed, within the individual zone files.

A specific zone file MUST define:
   Nest thermostat 'name long'
   IFTTT actions
   heating/cooling settings
   gpio toggles for heat, cool and fan

The main hvac.py has been reconfigured to be more efficient in the
while loops.

Signed-off-by: Mark Hatle <fray@kernel.crashing.org>