marksroom: Add a new window air conditioner and config zone
A new AEC08LX (8000 BTU) Window Air Conditioner was installed and the system
was configured appropriately.

It was observed that a -2F offset was not enough to cool the broader upstairs.
The part of the room closest to the air conditioner would often get to -2F
of the desired temp before the rest of the room.  Moving to -3F seems to have
resolved this issue.  The +2F offset seems to be adequate for the 'eco' (off)
mode.

At the same time it was noticed that the zone was missing the fan functions.
Fix this issue, so that we can (eventually) use IFTTT fans to help
control airflow.

The Zone also overrides the cooling functions to adjust the fan speed
based on the time of day.  Typically it's better to leave everything in
automatic mode, but in a bedroom that can be noisy.

Signed-off-by: Mark Hatle <fray@kernel.crashing.org>