zone: Add basic support hot heat-cool and eco modes
The basic support simply takes the high range of temps for the cooling
and will use that instead of a single target temp -- as this is not valid
in the heat-cool or eco mode.

The code and related temp settings will still need some adjustment
for proper high/low temp modes, but this should be a reasonable workaround
for cooling season.

Signed-off-by: Mark Hatle <fray@kernel.crashing.org>