zone: Fix a few misc issues
When we first start the app we were configuring various aspects of the
system, and then checking the nest and gpios and such, and immediately
changing them.  The IFTTT service sometimes misses temp setting events
if they come too quickly.

We also wait until the nest data has been initialized before starting to
process the GPIOs.  This does pose an issue if the nest service is down at
HVAC startup, but this should hopefully be rare.

Signed-off-by: Mark Hatle <fray@kernel.crashing.org>