zone: Add a new zone 'catroom' (not a real zone)
Living room has been upgraded from a Nest v1 to a v3.  A second window
air conditioner has been installed in the cat room, adjacent to the living
room zone.

This new air conditioner will be treated as a stage 2 cooling appliance for
the living room, but instead of implementing a special stage 2 -- we instead
add a second zone that triggers off the same nest as the living room zone,
but a different gpio.

Being a stage 2 cooling, it should only be used when necessary for additional
cooling or if the main system can not keep up.

Signed-off-by: Mark Hatle <fray@kernel.crashing.org>