hvac.py: Attempt to monitor and restart threads
Move the initial thread starting into the while loop.  This should
make it easier to manage thread creation in the future, as all of
the calls are now grouped.

Use this came code path to restart any dead threads.  Presumably they
died due to an exception of some kinds.

Recently there have been some issues with the Nest API connection
failing and nothing notices for a while.

Signed-off-by: Mark Hatle <fray@kernel.crashing.org>